arbitrary-type immediates
This is a first pass through converting immediates to arbitrary type.
test_internal passes, but nothing else is tested yet.

There are three IR nodes now: IntImm, UIntImm, and FloatImm. Their type
is meaningful. The immediate is stored as a int64_t, uint64_t, or
double.


Former-commit-id: 3f0f43c527191734cad2d0fad818977bf8ff71c9

+++ src/Deinterleave.cpp
@@ -552,14 +552,14 @@ class Interleaver : public IRMutator {
             // It's not a store of a ramp index.
             if (!r0) goto fail;
 
-            const int *stride_ptr = as_const_int(r0->stride);
+            const int64_t *stride_ptr = as_const_int(r0->stride);
 
             // The stride isn't a constant or is <= 0
             if (!stride_ptr || *stride_ptr < 1) goto fail;
 
-            const int stride = *stride_ptr;
+            const int64_t stride = *stride_ptr;
             const int width = r0->width;
-            const int expected_stores = stride == 1 ? width : stride;
+            const int64_t expected_stores = stride == 1 ? width : stride;
 
             // Collect the rest of the stores.
             std::vector<Store> stores;
@@ -593,7 +593,7 @@ class Interleaver : public IRMutator {
                 if (ri->width != width) goto fail;
 
                 Expr diff = simplify(ri->base - r0->base);
-                const int *offs = as_const_int(diff);
+                const int64_t *offs = as_const_int(diff);
 
                 // Difference between bases is not constant.
                 if (!offs) goto fail;
